Remove hugopl/version_from_shard dependency.
This shard was archived by me, the macro above does almost the same.

It get the version from shards.yml and add a build to it with a short
git hash or "unknown" if it wasn't built from a git repository (e.g. tarballs).

So version 0.15.0 is now show as 0.15.0+f8aa5c6.